feat(version): add version utility functions and refactor version comparison logic
