+++ apps/engineering-wizard/src/components/shared/ProcessInfoBar.svelte
@@ -76,7 +76,7 @@
     display: flex;
     flex-direction: row;
     gap: 12px;
-    align-items: center;
+    align-items: flex-start;
     padding: 12px 0;
     background: var(--white, #fff);
   }
