+++ libs/shared/src/index.ts
@@ -4,3 +4,4 @@ export * from './lib/PluginGroup';
 export * from './lib/LocalStoredPlugin';
 export * from './lib/CoMPASPlugin';
 export * from './lib/XPathValidation';
+export * from './lib/version.utils';
